Move serde into own file
