+++ algebra/primefield/Cargo.toml
@@ -38,6 +38,7 @@ quickcheck_macros = "0.9.1"
 rand = "0.7.2"
 rand_xoshiro = "0.4.0"
 raw-cpuid = "7.0.3"
+serde_json = "1.0"
 structopt = "0.3"
 zkp-criterion-utils = { version = "0.1.0", path = "../../utils/criterion-utils" }
 zkp-logging-allocator = { version = "0.1.0", path = "../../utils/logging-allocator" }
